+++ x/evm/keeper/keeper.go
@@ -78,6 +78,10 @@ type Keeper struct {
 	evmConstructor evm.Constructor
 	// Legacy subspace
 	ss paramstypes.Subspace
+
+	// a set of store keys that should cover all the precompile use cases,
+	// or ideally just pass the application's all stores.
+	keys map[string]*storetypes.KVStoreKey
 }
 
 // NewKeeper generates new evm module keeper
@@ -93,6 +97,7 @@ func NewKeeper(
 	evmConstructor evm.Constructor,
 	tracer string,
 	ss paramstypes.Subspace,
+	keys map[string]*storetypes.KVStoreKey,
 ) *Keeper {
 	// ensure evm module account is set
 	if addr := ak.GetModuleAddress(types.ModuleName); addr == nil {
@@ -118,7 +123,12 @@ func NewKeeper(
 		evmConstructor:    evmConstructor,
 		tracer:            tracer,
 		ss:                ss,
+		keys:              keys,
 	}
+}
+
+func (k Keeper) StoreKeys() map[string]*storetypes.KVStoreKey {
+	return k.keys
 }
 
 // Logger returns a module-specific logger.